Refactor Private Implementation
- Made _client implementation private.
- Made _compat implementation private.
- Made _config implementation private.
- Made _utils implementation private.
- Made _vendor implementation private.
- Made _factory implementation private.
- Refactored linting configurations.
