+++ crux/models/dataset.py
@@ -12,17 +12,17 @@ from typing import (  # noqa: F401 pylint: disable=unused-import
     Union,
 )
 
+from crux._utils import split_posixpath_filename_dirpath
 from crux.exceptions import CruxAPIError, CruxClientError, CruxResourceNotFoundError
-from crux.models.factory import get_resource_object
+from crux.models._factory import get_resource_object
 from crux.models.file import File
 from crux.models.folder import Folder
 from crux.models.job import LoadJob, StitchJob
 from crux.models.label import Label
 from crux.models.model import CruxModel
 from crux.models.query import Query
-from crux.models.resource import Resource
+from crux.models.resource import MediaType, Resource
 from crux.models.table import Table
-from crux.utils import MediaType, split_posixpath_filename_dirpath
 
 
 class Dataset(CruxModel):
@@ -53,7 +53,7 @@ class Dataset(CruxModel):
             website (str): Dataset website. Defaults to None.
             created_at (str): Dataset created. Defaults to None.
             modified_at (str): Dataset Modified. Defaults to None.
-            connection (crux.client.CruxClient): Connection Object. Defaults to None.
+            connection (crux._client.CruxClient): Connection Object. Defaults to None.
             raw_response (dict): Response Content. Defaults to None.
             tags (:obj:`list` of :obj:`str`): List of tags to be applied to dataset.
                 Defaults to None.
@@ -366,7 +366,7 @@ class Dataset(CruxModel):
             # hence raising the 404 error from the Python client
             raise CruxResourceNotFoundError({"statusCode": 404, "name": resource_name})
 
-    def resource_exists(self, path):
+    def _resource_exists(self, path):
         # type: (str) -> bool
         """Checks the existence of resource.
 
@@ -1139,7 +1139,7 @@ class Dataset(CruxModel):
         if isinstance(destination_resource, File):
             destination_file_object = destination_resource
         elif isinstance(destination_resource, str):
-            if self.resource_exists(path=destination_resource):
+            if self._resource_exists(path=destination_resource):
                 destination_file_object = self._get_resource(
                     path=destination_resource, model=File
                 )
