+++ crux/models/table.py
@@ -2,9 +2,9 @@
 
 from typing import Any, Dict  # noqa: F401 pylint: disable=unused-import
 
-from crux.compat import unicode
+from crux._compat import unicode
+from crux._utils import DEFAULT_CHUNK_SIZE
 from crux.models.resource import Resource
-from crux.utils import DEFAULT_CHUNK_SIZE
 
 
 class Table(Resource):
