+++ mypy.ini
@@ -1,5 +1,5 @@
 [mypy]
 python_version = 2.7
 
-[mypy-crux.vendor.*]
+[mypy-crux._vendor.*]
 ignore_errors = True
